CatUserBot v3.2.0 ( Major update ) (#308)
✚ Additions:
☞ Added VC support:
 • Required Var:
     VCMODE = True
     VC_SESSION = New Session (Optional)
• Required Buildpack:
     heroku-buildpack-nodejs --> https://github.com/rahulps1000/heroku-buildpack-nodejs
☞ Redesigned InlineQuery:
• Added Spotify, FileManager, CmdSearch.
• Added Multiple user support for Secret & Troll.

☞ Added New Chatbot (using Kuki):
• RANDOM_STUFF_API_KEY var is removed (no longer needed).

✚ Improvements:
☞ App: Redesigned & Fixed error.
☞ Gdrive: Added flag [-t] in [.glist] & post in telegraph for long list.
☞ Groupdata: Added [.grpstats]
☞ Heroku: Added [.buildpack]
☞ Logo: Added flag [-f] in [.logo] & Added random backgrounds (Do [.lbg] to get)
☞ Notebook: Added flag [-f] in [.write]
☞ Pastebin: Added [.rayso]
☞ Songs: Added flag [s] in [.shazam]
☞ Spam: Added [.rspam]
☞ Spotify: Added inline query & changed bot to @CatMusicRobot for [.now]
☞ Stats: Added flag [p] in [.stat] to get public links & changed to multiple texts for long message instead of file.
☞ Vps: Added branch switch support, can switch branch like heroku now on vps also.
☞ Whois: Added Is_Premium

✚ Bug Fixes:
☞ Android: [.device] - Error on Uppercase.
☞ Anilist: [.mal] - Error on Response.
☞ Carbon: Webdriver Error (Removed .krb).
☞ Dictionary: [.meaning] - Reformatted using Api (Removed PyDictonary).
☞ Fake: [.prankpromote] - Error on promote, if other than owner.
☞ Ffmpeg: - Error on document files.
☞ FileConverts: [.gif / .vtog] - Large gif size.
☞ Lyrics: - Error on Vps.
☞ Ocr: - Error on getting image from file.
☞ Quotly: [.q] - Improved font.
☞ Spam: [.spspam] - Error for Premium sticker.
☞ Stickers: - Error on kanging video & animated stickers.
☞ Ytdl: [.insta] - Error on getting videos.

+++ userbot/helpers/qhelper.py
@@ -41,44 +41,50 @@ logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
 
-def file_check():
+def file_check(re=True, me=True, mo=True, it=True, fa=True, sp=True, go=False):
     regular = "./temp/Roboto-Regular.ttf"
     medium = "./temp/Roboto-Medium.ttf"
     mono = "./temp/DroidSansMono.ttf"
     italic = "./temp/Roboto-Italic.ttf"
     fallback = "./temp/Quivira.otf"
     special = "./temp/ArialUnicodeMS.ttf"
+    google = "./temp/GoogleSans-Medium.ttf"
     if not os.path.isdir("./temp/"):
-        os.mkdir("./temp/", 0o755)
-    if not os.path.exists(regular):
+        os.mkdir("./temp/")
+    if re and not os.path.exists(regular):
         urllib.request.urlretrieve(
             "https://github.com/erenmetesar/modules-repo/raw/master/Roboto-Regular.ttf",
             regular,
         )
-    if not os.path.exists(medium):
+    if me and not os.path.exists(medium):
         urllib.request.urlretrieve(
             "https://github.com/erenmetesar/modules-repo/raw/master/Roboto-Medium.ttf",
             medium,
         )
-    if not os.path.exists(mono):
+    if mo and not os.path.exists(mono):
         urllib.request.urlretrieve(
             "https://github.com/erenmetesar/modules-repo/raw/master/DroidSansMono.ttf",
             mono,
         )
-    if not os.path.exists(italic):
+    if it and not os.path.exists(italic):
         urllib.request.urlretrieve(
             "https://github.com/erenmetesar/modules-repo/raw/master/Roboto-Italic.ttf",
             italic,
         )
-    if not os.path.exists(fallback):
+    if fa and not os.path.exists(fallback):
         urllib.request.urlretrieve(
             "https://github.com/erenmetesar/modules-repo/raw/master/Quivira.otf",
             fallback,
         )
-    if not os.path.exists(special):
+    if sp and not os.path.exists(special):
         urllib.request.urlretrieve(
             "https://github.com/TgCatUB/CatUserbot-Resources/blob/master/Resources/Spotify/ArialUnicodeMS.ttf?raw=true",
             special,
+        )
+    if go and not os.path.exists(google):
+        urllib.request.urlretrieve(
+            "https://github.com/TgCatUB/CatUserbot-Resources/blob/master/Resources/Spotify/GoogleSans-Medium.ttf?raw=true",
+            google,
         )
